renaming files and modules
