+++ .github/workflows/valgrind.yml
@@ -37,6 +37,7 @@ jobs:
       # Execute tests defined by the CMake configuration.
       # See https://cmake.org/cmake/help/latest/manual/ctest.1.html for more detail
       run: 
-        cd ${{github.workspace}}/build/tests/ArrayTestSuite && valgrind -s -v --error-exitcode=1 --leak-check=full ./ArrayTestSuite &&
-        cd ${{github.workspace}}/build/tests/StackTestSuite && valgrind -s -v --error-exitcode=1 --leak-check=full ./StackTestSuite &&
-        cd ${{github.workspace}}/build/tests/VectorTestSuite && valgrind -s -v --error-exitcode=1 --leak-check=full ./VectorTestSuite
+        cd ${{github.workspace}}/build/tests/array_test_suite && valgrind -s -v --error-exitcode=1 --leak-check=full ./array_test_suite &&
+        cd ${{github.workspace}}/build/tests/stack_test_suite && valgrind -s -v --error-exitcode=1 --leak-check=full ./stack_test_suite &&
+        cd ${{github.workspace}}/build/tests/vector_test_suite && valgrind -s -v --error-exitcode=1 --leak-check=full ./vector_test_suite &&
+        cd ${{github.workspace}}/build/tests/list_test_suite && valgrind -s -v --error-exitcode=1 --leak-check=full ./list_test_suite
