+++ scripts/run_all_unit_tests.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # TODO: redesign this part to run all tests without adding new element of calling here
-./ArrayTestSuite/ArrayTestSuite;
-./ListTestSuite/ListTestSuite;
-./StackTestSuite/StackTestSuite;
-./VectorTestSuite/VectorTestSuite;
+./array_test_suite/array_test_suite;
+./list_test_suite/list_test_suite;
+./stack_test_suite/stack_test_suite;
+./vector_test_suite/vector_test_suite;
