+++ src/CMakeLists.txt
@@ -1,5 +1,3 @@
-cmake_minimum_required(VERSION 3.24)
-
 add_executable(
     ${PROJECT_NAME}
         main.cpp)
