+++ tests/CMakeLists.txt
@@ -1,5 +1,3 @@
-cmake_minimum_required(VERSION 3.24)
-
 find_package(Git REQUIRED)
 
 set(GTEST_ROOT ${CMAKE_BINARY_DIR}/googletest)
@@ -18,10 +16,10 @@ link_directories(${GTEST_ROOT}/lib)
 
 set(GTest_DIR "${GTEST_ROOT}/lib/cmake/GTest")
 
-add_subdirectory(ArrayTestSuite)
-add_subdirectory(ListTestSuite)
-add_subdirectory(StackTestSuite)
-add_subdirectory(VectorTestSuite)
+add_subdirectory(array_test_suite)
+add_subdirectory(list_test_suite)
+add_subdirectory(stack_test_suite)
+add_subdirectory(vector_test_suite)
 
 # Copy unit test runnning script
 file(
