+++ tests/array_test_suite/empty_array_test_suite.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "Array.hpp"
+#include "array.hpp"
 
 namespace {
