+++ tests/vector_test_suite/CMakeLists.txt
@@ -1,10 +1,8 @@
-cmake_minimum_required(VERSION 3.24)
-
-set(STACK_TEST_SUITE_NAME "StackTestSuite")
+set(STACK_TEST_SUITE_NAME "vector_test_suite")
 add_executable(
     ${STACK_TEST_SUITE_NAME}
-        StackTestSuite.cpp
-        EmptyStackTestSuite.cpp)
+        vector_test_suite.cpp
+        empty_vector_test_suite.cpp)
 
 target_include_directories(
     ${STACK_TEST_SUITE_NAME}
