+++ tests/vector_test_suite/empty_vector_test_suite.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "Vector.hpp"
+#include "vector.hpp"
 
 namespace
 {
