+++ tests/vector_test_suite/vector_test_suite.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "Vector.hpp"
+#include "vector.hpp"
 
 namespace
 {
